requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
httpx[http2]>=0.24.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
requests>=2.28.0
tidalapi=0.8.8
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
websockets>=11.0